Bug: Fixed double to BigDecimal data types

+++ src/main/java/br/com/alelo/consumer/consumerpat/entity/Extract.java
@@ -4,6 +4,8 @@ import lombok.Data;
 
 import javax.persistence.Entity;
 import javax.persistence.Id;
+
+import java.math.BigDecimal;
 import java.util.Date;
 
 
@@ -18,10 +20,10 @@ public class Extract {
     String productDescription;
     Date dateBuy;
     int cardNumber;
-    double value;
+    BigDecimal value;
 
 
-    public Extract(int id, int establishmentNameId, String establishmentName, String productDescription, Date dateBuy, int cardNumber, double value) {
+    public Extract(int id, int establishmentNameId, String establishmentName, String productDescription, Date dateBuy, int cardNumber, BigDecimal value) {
         this.id = id;
         this.establishmentNameId = establishmentNameId;
         this.establishmentName = establishmentName;
@@ -31,14 +33,14 @@ public class Extract {
         this.value = value;
     }
 
-    public Extract( String productDescription, Date dateBuy, int cardNumber, double value) {
+    public Extract( String productDescription, Date dateBuy, int cardNumber, BigDecimal value) {
         this.productDescription = productDescription;
         this.dateBuy = dateBuy;
         this.cardNumber = cardNumber;
         this.value = value;
     }
 
-    public Extract(String establishmentName, String productDescription, Date dateBuy, int cardNumber, double value) {
+    public Extract(String establishmentName, String productDescription, Date dateBuy, int cardNumber, BigDecimal value) {
         this.establishmentNameId = establishmentNameId;
         this.establishmentName = establishmentName;
         this.productDescription = productDescription;
